Add core express server file
This commit adds the basic express server config to the backend to allow for development. It allows for the port to be specified through the environment PORT (often used by hosts to specify what port to listen on. If this is not set it defaults to 8080. This commit also adds any file under the build folder as a static resource, which allows for any of the build files to be accessible through url access.