Add basic routing for API
This commit adds the basic API routing for the application. It exposes the following endpoints /api/profile/info and /api/profile/repos for use in the future. If the user tries to navigate to any other endpoint under /api or uses an incorrect verb the api error middleware will return a 404 not found response.