Add winston logging utility
This commit add a winston logger to be used within the application. It involves two different transports, one for the hourly file rotation and one that outputs to the console. 

All logs are outputted in the following format: '${info.timestamp} [${info.level}]: ${info.message}' this allows for easier navigation of logs. To assist with keeping files short the max size of the file is 20m. After holding 3 days of logs it will archive older ones and zip them up. A 'log' folder is also specified that all logs will be outputted to. 

To reduce log sizes in production mode, we also specify to only show debug logs in a non-production environment.